@@ -1,7 +1,7 @@
 package com.jantursky.debugger.adapters;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -9,7 +9,6 @@ import android.widget.TextView;
 
 import com.jantursky.debugger.R;
 import com.jantursky.debugger.components.DebuggerComponent;
-import com.jantursky.debugger.components.dbviewer.DbViewerComponent;
 import com.jantursky.debugger.listeners.DebuggerComponentItemListener;
 import com.jantursky.debugger.utils.AnnotationUtils;
 
