@@ -4,14 +4,13 @@ import android.app.Activity;
 import android.app.Fragment;
 import android.content.Context;
 import android.graphics.Color;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.core.content.ContextCompat;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.view.ViewGroup;
 
 import com.jantursky.debugger.adapters.DebuggerComponentAdapter;
-import com.jantursky.debugger.components.dbviewer.DbViewerComponent;
 import com.jantursky.debugger.listeners.DebuggerComponentItemListener;
 import com.jantursky.debugger.interfaces.ComponentListener;
 import com.jantursky.debugger.utils.RecyclerViewUtils;
