@@ -1,7 +1,7 @@
 package com.jantursky.debugger.components.dbviewer.helpers;
 
 import android.graphics.Rect;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 public class ItemDecorationGrid extends RecyclerView.ItemDecoration {
