@@ -1,6 +1,6 @@
 package com.jantursky.debugger.components.sharedpreferencesviewer.annotations;
 
-import android.support.annotation.IntDef;
+import androidx.annotation.IntDef;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
